+++ src/textual/screen.py
@@ -49,6 +49,8 @@ from .widgets._toast import ToastRack
 if TYPE_CHECKING:
     from typing_extensions import Final
 
+    from .command_palette import CommandSource
+
     # Unused & ignored imports are needed for the docs to link to these objects:
     from .errors import NoWidget  # type: ignore  # noqa: F401
     from .message_pump import MessagePump
@@ -154,6 +156,9 @@ class Screen(Generic[ScreenResultType], Widget):
     """Screen sub-title to override [the app sub-title][textual.app.App.sub_title]."""
     title: Reactive[str | None] = Reactive(None, compute=False)
     """Screen title to override [the app title][textual.app.App.title]."""
+
+    COMMAND_SOURCES: ClassVar[set[type[CommandSource]]] = set()
+    """The [command sources](/api/command_palette/) for the screen."""
 
     BINDINGS = [
         Binding("tab", "focus_next", "Focus Next", show=False),